refactor: problem::solve() wraps stateful solution::ProxySolution
thread: https://github.com/cfm/proxy-solver-api/pull/9#discussion_r1312868556
